requirements-piper.txt
piper-tts==1.2.0
numpy
